+++ flask_via/examples/blueprints/baz/views.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+
+"""
+flask_via.examples.blueprints.baz.views
+=======================================
+
+A blueprint ``Flask-Via`` example Flask application.
+"""
+
+from flask import render_template
+
+
+def baz():
+    return render_template('baz.html')